Preliminary fix for detecting the remote IP through the nginx proxy.
This probably works (it does in EmergentMUD) but it needs to be refactored to be testable. The static call into RequestContextHolder is no bueno and would need PowerMock to deal with it. I don't have time right now but the next step will be to wrap that into some kind of non-static thing so Mockito can deal with it.