Turned off open-in-view
The default setting for open-in-view seems to change the behavior of JPA from what it was in earlier versions of Spring. I'm just going to turn it off for now, which seems to restore the behavior I expected to see. It seems like it could be pretty useful since it would remove the need to call save() explicitly after editing a managed entity, but I want to read up some more about it to learn what the implications and benefits really are. I also don't want to engage in a major refactor of anything at 2 AM.