WIP - Refining the way event participants are managed
* Removed the multi-select box from create and edit screens
* You must create the event without participants, then edit to add them
* Participant emails are now required to be unique in the database
* Edit screen allows adding each participant by email address
* Remove doesn't work yet, that'll be the next PR

Event edits no longer include changes to participants. Each participant change will be a separate REST call and happen immediately. When there are many participants in the database and many participating in an event, this strategy will be much more scalable than the multi-select box.

This would be a good argument for moving to something like Angular or React because we will have problems when several people are changing participants on an event at the same time. Right now, if someone else changes the list, your view of it won't update until you reload the page. Some of those more modern frameworks provide real time updates when the underlying data changes, so you'd be able to see changes from other users immediately as they happen. That's a big lift that I'll save for post-MVP.