Award data model
Award is the JPA representation of the event_participants join table, which is getting a new "asset_id" field. The new field will allow us to assign an asset to a person's participation in an event, which is a fancy way of saying that they won a prize for being part of a contest.