Assign awards
The select boxes now work to pick an award for a participant within an event. There is a unique constraint in the database that prevents assigning the same asset to more than one person at a time. The error message leaves something to be desired, but I'm leaving that for another day.